Create hook directory
