+++ tgm/hooks/hooks.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import is_dataclass
-from typing import Any, List, Protocol, Set, runtime_checkable
+from typing import Any, List, Set
 
 import numpy as np
 import torch
@@ -9,94 +9,7 @@ import torch
 from tgm import DGBatch, DGraph
 from tgm._storage import DGSliceTracker
 from tgm.constants import PADDED_NODE_ID
-
-
-@runtime_checkable
-class DGHook(Protocol):
-    r"""The behaviours to be executed on a DGraph before materializing."""
-
-    requires: Set[str]
-    produces: Set[str]
-    has_state: bool
-
-    def __call__(self, dg: DGraph, batch: DGBatch) -> DGBatch: ...
-
-    def reset_state(self) -> None: ...
-
-
-class StatelessHook:
-    requires: Set[str] = set()
-    produces: Set[str] = set()
-    has_state: bool = False
-
-    def __call__(self, dg: DGraph, batch: DGBatch) -> DGBatch:
-        raise NotImplementedError
-
-    def reset_state(self) -> None:
-        pass
-
-
-class StatefulHook:
-    requires: Set[str] = set()
-    produces: Set[str] = set()
-    has_state: bool = True
-
-
-class HookManager:
-    def __init__(self, dg: DGraph, hooks: List[DGHook]) -> None:
-        if not isinstance(hooks, list):
-            raise TypeError(f'Invalid hook type: {type(hooks)}')
-        bad_hook_names = [type(h).__name__ for h in hooks if not isinstance(h, DGHook)]
-        if len(bad_hook_names):
-            raise TypeError(
-                f'These hooks do not correctly implement the DGHook protocol: {bad_hook_names}, '
-                'ensure there is a __call__(self, dg: DGraph, batch: DGBatch) -> DGBatch implemented'
-            )
-
-        # Implicitly add dedup hook after all user-defined hooks and before device transfer
-        hooks.append(DeduplicationHook())
-
-        if dg.device.type != 'cpu':
-            hooks.append(PinMemoryHook())
-            hooks.append(DeviceTransferHook(dg.device))
-
-        self.hooks = hooks
-        self._validate_hook_dependencies()
-
-    def reset_state(self) -> None:
-        for hook in self.hooks:
-            hook.reset_state()
-
-    @classmethod
-    def from_any(
-        cls, dg: DGraph, hook_like: HookManager | DGHook | List[DGHook] | None
-    ) -> HookManager:
-        if isinstance(hook_like, cls):
-            return hook_like
-        elif hook_like is None:
-            return cls(dg, hooks=[])
-        elif isinstance(hook_like, DGHook):
-            return cls(dg, hooks=[hook_like])
-        elif isinstance(hook_like, list):
-            return cls(dg, hooks=hook_like)
-        else:
-            raise TypeError(f'Invalid hook type: {type(hook_like)}')
-
-    def __call__(self, dg: DGraph) -> DGBatch:
-        batch = dg.materialize()
-        for hook in self.hooks:
-            batch = hook(dg, batch)
-        return batch
-
-    def _validate_hook_dependencies(self) -> None:
-        produced: Set[str] = set()
-        for hook in self.hooks:
-            missing = hook.requires - produced
-            if missing:
-                raise ValueError(
-                    f'{hook.__class__.__name__} is missing required fields: {missing}'
-                )
-            produced |= hook.produces
+from tgm.hooks import StatefulHook, StatelessHook
 
 
 class PinMemoryHook(StatelessHook):
